Ideja s long arrayem

+++ src/main/java/hr/fer/zemris/optjava/dz3/IFunction.java
@@ -2,5 +2,5 @@ package hr.fer.zemris.optjava.dz3;
 
 public interface IFunction {
 
-	double valueAt(double[] point);
+	double valueAt(long[] point);
 }
